+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 
 ## [1.16.0]
 ### Modified
+
 - squeeze pandas dataframe on csv_read
+- load_cache now public method; option to load specific remote cache with tag arg
 
 ## [1.15.0]